+++ src/components/WeatherDisplay/WeatherCard/WeatherCard.module.css
@@ -1,25 +1,3 @@
-.weather {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    font-size: 1.2em;
-    border-radius: 1em;
-    background-color: #fff;
-    color: black;
-    width: 20em;
-    height: 20em;
-    box-shadow: 0 0 1em #ccc;
-
-
-}
-
-.weather__city {
-    display: flex;
-    margin: 1em;
-    align-items: center;
-}
-
 .weather__condition__icon {
     font-size: 1.3em;
     display: flex;
@@ -27,4 +5,3 @@
     justify-content: space-evenly;
     align-items: center;
 }
-
