+++ src/components/WeatherDisplay/WeatherDisplay.module.css
@@ -1,0 +1,18 @@
+.weather__cards {
+    display: grid;
+    grid-template-columns: 1fr 2fr 1fr;
+    align-items: center;
+    justify-content: center;
+
+}
+
+.weather__cards article {
+    margin: 0.5em;
+    padding: 0.5em;
+    font-size: 1.2em;
+    border-radius: 1em;
+    background-color: #fff;
+    color: black;
+    box-shadow: 0 0 0.5em #ccc;
+}
+
